Add near handler
This handler returns any routes near the specified point. As is, it's
quite inefficient because it queries for stops near the point, and then
for each returned stop, asks for the route data of routes that pass by.

In the future, we'll probably want to either cache this or just have the
GTFS database available.